Use total population instead of population in housing units
We use the population as a denominator for the racial makeup of a block group. Previously we had used the population in occupied housing units. This led to some block groups near UC Berkeley to show 600% of the population as Asian because university dorms are not considered occupied housing units, but rather group quarters.